+++ src/util/error.ts
@@ -3,6 +3,7 @@ import config from '../fabric/config'
 import { logger } from './logger'
 
 class BdkError extends Error {}
+class BdkWarn extends Error {}
 
 export class ProcessError extends BdkError {}
 export class ParamsError extends BdkError {}
@@ -10,6 +11,13 @@ export class DockerError extends BdkError {}
 export class BackupError extends BdkError {}
 export class TimeLimitError extends BdkError {}
 export class PathError extends BdkError {}
+export class SolcError extends BdkError {}
+export class SolcWarn extends BdkWarn {}
+export class NotFoundWarn extends BdkWarn {}
+export class FileWriteError extends BdkError {}
+export class DeployError extends BdkError {}
+export class DataError extends BdkError {}
+
 export class FabricContainerError extends BdkError {
   public stdout: string
   constructor (message: string, stdout: string) {
@@ -17,6 +25,7 @@ export class FabricContainerError extends BdkError {
     this.stdout = stdout
   }
 }
+
 export class EthContainerError extends BdkError {
   public stdout: string
   constructor (message: string, stdout: string) {
@@ -39,6 +48,8 @@ export const errorHandler = (err: Error) => {
     logger.error(err.message)
   } else if (err instanceof BdkError) {
     logger.error(err.message)
+  } else if (err instanceof BdkWarn) {
+    logger.warn(err.message)
   } else {
     logger.error('Unexpected error.\n')
     logger.error(err.message)