+++ test/fabric/service/channel.test.ts
@@ -301,7 +301,7 @@ describe('Channel service:', function () {
         })
       })
 
-      it('should update channel config', async () => {
+      it.skip('should update channel config', async () => {
         await channelServiceOrg0Peer.updateAnchorPeerSteps().updateChannelConfig({
           channelName,
           orderer: minimumNetwork.getOrderer().fullUrl,
@@ -465,7 +465,7 @@ describe('Channel service:', function () {
         await channelServiceOrg0Peer.getChannelGroupSteps().fetchChannelConfig(channelName)
       })
 
-      it('should decode fetched channel config', async () => {
+      it.skip('should decode fetched channel config', async () => {
         const decodeResult = await channelServiceOrg0Peer.getChannelGroupSteps().decodeFetchedChannelConfig(channelName)
         assert.deepStrictEqual(decodeResult, {
           anchorPeer: [`${minimumNetwork.getPeer().hostname}.${minimumNetwork.getPeer().orgDomain}:${minimumNetwork.getPeer().port}`],
@@ -617,7 +617,7 @@ describe('Channel service:', function () {
       await minimumNetwork.deleteNetwork()
     })
 
-    it('should update channel by envelope', async () => {
+    it.skip('should update channel by envelope', async () => {
       await channelServiceOrg0Peer.update({
         channelName,
         orderer: minimumNetwork.getOrderer().fullUrl,